Version 2 of the CollaborativeContext implementation.
This version separates out the "shadows" as a referencable Context in its own right,
which makes things slightly more robust, reduces the number of hashtable lookups,
and allows for (eventual) parallel processing of the CollaborativeContexts.

Don't try with multiple ranks, it still won't work, it just won't complain very hard.