Hide our file descriptors from the application so they don't get closed on us (or worse).
Applications like Firefox (#402) close all the open file descriptors as a security measure, and seem to be sensitive to the
file descriptor numbers. When we're opening files in the middle of that things tend to error in magnificent ways. This
prototype fix moves our descriptors to a location above the hard limit on the file descriptor number, by placing it just
below the limit and then lowering the limit accordingly. Since lowering the hard limit is an irreversible change (except for
root), this effectively ensures our file descriptors are untouchable from the application.

Technically the file descriptors are still visible through /proc/self/fd, but Firefox considers "above the limit" to be
hidden for compatibility with Valgrind (see https://hg.mozilla.org/mozilla-central/file/6fd4bb500d425c406c1b52f66e5b195b20ae5e0a/ipc/chromium/src/base/process_util_posix.cc#l164).

This is a rough prototype, there may be issues with other frameworks using the same trick or if the application saves the
limit early. A more robust approach would reserve a set of files (and descriptors) for internal use very early and rename as
needed, which would require a one-file-per-process approach to trace files and a non-pipe communication link with fnbounds.